refactor(binder): split declaration flow helpers
