+++ crates/tsz-binder/src/binding/mod.rs
@@ -3,7 +3,10 @@
 //! - `declaration.rs`: declaration binding, accessors, and flow graph construction.
 //! - `validation.rs`: post-binding validation, lib symbol diagnostics, and resolution statistics.
 
+mod accessors_flow;
 pub(crate) mod declaration;
+mod expression_flow;
+mod semantic_defs;
 pub(crate) mod stack_guard;
 mod validation;
